+++ src/suppressAuthInterceptor/suppressAuthInterceptor.ts
@@ -1,7 +1,7 @@
 import {ClassProvider, Injectable} from '@angular/core';
 import {HttpInterceptor, HTTP_INTERCEPTORS, HttpEvent, HttpHandler, HttpRequest} from '@angular/common/http';
-import {IgnoredInterceptorsService, AdditionalInfo, IgnoredInterceptorId} from '@anglr/common';
-import {Observable, ObservableInput, Observer} from 'rxjs';
+import {IGNORED_INTERCEPTORS} from '@anglr/common';
+import {Observable, ObservableInput} from 'rxjs';
 import {catchError} from 'rxjs/operators';
 
 /**
@@ -10,11 +10,6 @@ import {catchError} from 'rxjs/operators';
 @Injectable()
 export class SuppressAuthInterceptor implements HttpInterceptor
 {
-    //######################### constructors #########################
-    constructor(private _ignoredInterceptorsService: IgnoredInterceptorsService)
-    {
-    }
-
     //######################### public methods - implementation of HttpInterceptor #########################
 
     /**
@@ -22,15 +17,15 @@ export class SuppressAuthInterceptor implements HttpInterceptor
      * @param req - Request to be intercepted
      * @param next - Next middleware that can be called for next processing
      */
-    public intercept(req: HttpRequest<any> & AdditionalInfo<IgnoredInterceptorId>, next: HttpHandler): Observable<HttpEvent<any>>
+    public intercept(req: HttpRequest<unknown>, next: HttpHandler): Observable<HttpEvent<unknown>>
     {
         return next.handle(req).pipe(catchError((err) =>
         {
-            return Observable.create((observer: Observer<any>) =>
+            return new Observable(observer =>
             {
                 //client error, not response from server, or is ignored
                 if (err.error instanceof Error || 
-                    (this._ignoredInterceptorsService && this._ignoredInterceptorsService.isIgnored(SuppressAuthInterceptor, req.additionalInfo)))
+                    req.context.get(IGNORED_INTERCEPTORS).some(itm => itm == SuppressAuthInterceptor))
                 {
                     observer.error(err);
                     observer.complete();
@@ -49,7 +44,7 @@ export class SuppressAuthInterceptor implements HttpInterceptor
                 //other errors
                 observer.error(err);
                 observer.complete();
-            }) as ObservableInput<HttpEvent<any>>;
+            }) as ObservableInput<HttpEvent<unknown>>;
         }));
     }
 }