+++ src/testing/common/fakeAuthorize.directive.ts
@@ -15,7 +15,7 @@ export class FakeAuthorizeDirective
      * Name of permission that is requested for displaying element
      */
     @Input('authorize')
-    public permission: string | string[];
+    public permission: string | string[] = null!;
 
     /**
      * Indication that AND condition should be used instead of OR condition if multiple permissions are provided
@@ -30,7 +30,7 @@ export class FakeAuthorizeDirective
     public conditionString: boolean = false;
 
     //######################### constructor #########################
-    constructor(private _template: TemplateRef<any>,
+    constructor(private _template: TemplateRef<unknown>,
                 private _viewContainer: ViewContainerRef)
     {
         this._viewContainer.createEmbeddedView(this._template);